+++ src/compat/predicate/isSymbol.ts
